# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FTTDDWA Member Management API
  description: |
    REST API for the **Federation of Tamil Nadu Tent Dealers & Decorators Welfare Association**
    member management system.

    ## Authentication
    All protected endpoints require a valid JWT access token delivered as either:
    - **Cookie** `access_token` (set automatically on login — preferred for browser clients)
    - **Header** `Authorization: Bearer <token>` (for API/mobile clients)

    Access tokens expire in **15 minutes**. Use `POST /auth/refresh` to obtain a new one
    using the `refresh_token` HttpOnly cookie (valid 7 days).

    ## Rate Limits
    | Scope | Limit |
    |-------|-------|
    | Auth endpoints | 5 requests / 15 min per IP |
    | All API endpoints | 100 requests / 15 min per IP |

  version: "1.4.0"
  contact:
    name: FTTDDWA IT Support
    email: admin@fttddwa.org
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://portal.fttddwa.org/api
    description: Production
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── REUSABLE SCHEMAS ───────────────────────────────────────────────────────
  schemas:
    ApiSuccess:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error description"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email address"]
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            success: false
            message: "Unauthorized"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            success: false
            message: "Forbidden"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            success: false
            message: "Validation failed"
            errors:
              phone: ["Enter a valid 10-digit Indian mobile number"]
    RateLimit:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"

  # ─── REUSABLE PARAMETERS ────────────────────────────────────────────────────
...
security:
  - cookieAuth: []
  - bearerAuth: []

# ─── PATHS ────────────────────────────────────────────────────────────────────
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: |
        Authenticates a user with email and password.
        Sets two HttpOnly cookies: `access_token` (15 min) and `refresh_token` (7 days, path restricted to `/api/auth/refresh`).
        The response body does **not** contain user data — fetch `/auth/me` after login to get the user profile.

        Returns `429` after too many failed attempts from the same IP.
        Returns `401` after **10** consecutive wrong passwords — account is locked for 5 minutes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "admin@fttddwa.org"
                password:
                  type: string
                  format: password
                  example: "Admin@123456"
      responses:
        "200":
          description: Login successful — call `GET /auth/me` to retrieve user profile
          headers:
            Set-Cookie:
              description: "access_token (HttpOnly, 15 min) and refresh_token (HttpOnly, 7 days, path=/api/auth/refresh)"
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
              example:
                success: true
                message: "Login successful"
        "401":
          description: Invalid credentials or account locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          $ref: "#/components/responses/RateLimit"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Uses the `refresh_token` cookie to issue a new `access_token` and rotates the refresh token.
        The old refresh token is immediately invalidated.
      security: []
      responses:
        "200":
          description: Token refreshed
          headers:
            Set-Cookie:
              description: New access_token and refresh_token cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /members/register:
    post:
      tags: [Members]
      summary: Public member self-registration
      description: |
        Allows a prospective member to register without an admin account.
        - Returns `403` if `enableMemberRegistration` is disabled in App Settings.
        - Created member has status `INACTIVE` — admin must approve.
        - No authentication required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, address, district, taluk, phone, position]
              properties:
                name:
                  type: string
                  example: "Rajan Kumar"
                nameTamil:
                  type: string
                  example: "ராஜன் குமார்"
                position:
                  type: string
                  example: "Member"
                address:
                  type: string
                  example: "12 Main Street, Coimbatore"
                district:
                  type: string
                  example: "Coimbatore"
                taluk:
                  type: string
                  example: "Coimbatore North"
                state:
                  type: string
                  example: "Tamil Nadu"
                village:
                  type: string
                phone:
                  type: string
                  maxLength: 10
                  example: "9876543210"
                email:
                  type: string
                  format: email
                dateOfBirth:
                  type: string
                  format: date
                businessName:
                  type: string
      responses:
        "201":
          description: Registration submitted — pending admin approval
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          membershipId:
                            type: string
                            example: "FTTD260042"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          description: Member registration is currently disabled
        "409":
          description: Phone or email already registered
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiError"
                  - type: object
                    properties:
                      errors:
                        type: object
                        properties:
                          phone:
                            type: array
                            items:
                              type: string
                          email:
                            type: array
                            items:
                              type: string
...
  /upload:
    post:
      tags: [Upload]
      summary: Upload member photo
      description: |
        Uploads a passport-size member photo.
        - Accepted types: `image/jpeg`, `image/png`, `image/webp`
        - Maximum size: **5 MB**
        - Returns a URL path to be stored in the member's `photoUrl` field.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          url:
                            type: string
                            example: "/uploads/members/1714123456-abc123.jpg"
                          filename:
                            type: string
                            example: "1714123456-abc123.jpg"
        "400":
          description: Invalid file type or size exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── WHATSAPP ──────────────────────────────────────────────────────────────────
...
  /whatsapp/webhook:
    post:
      tags: [WhatsApp]
      summary: Delivery status webhook
      description: |
        **Public endpoint** — no authentication required.
        Called by the WhatsApp vendor when a message status changes (e.g., `delivered`, `read`).

        Verifies HMAC-SHA256 signature via `X-Webhook-Signature` header if `WHATSAPP_WEBHOOK_SECRET` is configured.
        Updates the corresponding log entry status. `read` status is normalized to `delivered`.
      security: []
      parameters:
        - name: X-Webhook-Signature
          in: header
          description: "HMAC-SHA256 of the raw request body, hex-encoded"
          required: false
          schema:
            type: string
            example: "sha256=abc123..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messageId, status]
              properties:
                messageId:
                  type: string
                  example: "msg_abc123"
                status:
                  type: string
                  enum: [sent, delivered, read, failed]
                  example: "delivered"
                timestamp:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
        "401":
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /whatsapp/cron/trigger:
    post:
      tags: [WhatsApp]
      summary: Trigger daily cron manually
      description: |
        Manually triggers the daily WhatsApp cron job (birthday/anniversary/event messages).
        Accepts either `X-Cron-Secret` header matching `WHATSAPP_CRON_SECRET` env var,
        or a valid JWT for a SUPER_ADMIN or ADMIN role.
        Returns counts of enqueued and skipped messages, or a `stoppedReason` if halted early (e.g., no vendor, wallet empty).
      parameters:
        - name: X-Cron-Secret
          in: header
          description: Internal cron secret (alternative to JWT auth)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Cron run result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          enqueued:
                            type: integer
                            example: 5
                          skipped:
                            type: integer
                            example: 2
                          processed:
                            type: integer
                            example: 7
                          failed:
                            type: integer
                            example: 0
                          durationMs:
                            type: integer
                            example: 312
                          errors:
                            type: array
                            items:
                              type: string
                          stoppedReason:
                            type: string
                            nullable: true
                            example: null
                          dlr:
                            type: object
                            nullable: true
                            description: "DLR poll results if status sync ran alongside cron"
                            properties:
                              checked:
                                type: integer
                              updated:
                                type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...